+++ Front/GESCOMPAH/src/app/features/setting/components/department/department.component.ts
@@ -7,9 +7,10 @@ import { Observable } from 'rxjs';
 import { MatDialog } from '@angular/material/dialog';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
-import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
 import { CommonModule } from '@angular/common';
 import { FormDialogComponent } from '../../../../shared/components/form-dialog/form-dialog.component';
+import { SweetAlertService } from '../../../../shared/Services/sweet-alert/sweet-alert.service'
+
 
 @Component({
   selector: 'app-department',
@@ -19,7 +20,6 @@ import { FormDialogComponent } from '../../../../shared/components/form-dialog/f
     GenericTableComponent,
     MatButtonModule,
     MatIconModule,
-    MatSnackBarModule
   ],
   templateUrl: './department.component.html',
   styleUrl: './department.component.css',
@@ -36,7 +36,7 @@ export class DepartmentComponent implements OnInit {
   constructor(
     private store: DepartmentStore,
     private dialog: MatDialog,
-    private snackbar: MatSnackBar
+    private sweetAlert: SweetAlertService
   ) {
     this.departments$ = this.store.departments$;
   }
@@ -58,7 +58,7 @@ export class DepartmentComponent implements OnInit {
       if (result) {
         this.store.create(result).subscribe({
           next: () => {
-            this.snackbar.open('Departamento creado exitosamente', 'Cerrar', { duration: 2000 });
+            this.sweetAlert.showNotification('Éxito', 'Departamento creado exitosamente', 'success');
           }
         });
       }
@@ -78,7 +78,7 @@ export class DepartmentComponent implements OnInit {
       if (result) {
         this.store.update(result).subscribe({
           next: () => {
-            this.snackbar.open('Departamento actualizado exitosamente', 'Cerrar', { duration: 2000 });
+            this.sweetAlert.showNotification('Éxito', 'Departamento actualizado exitosamente', 'success');
           }
         });
       }
@@ -87,9 +87,13 @@ export class DepartmentComponent implements OnInit {
 
   handleDelete(row: DepartmentModel): void {
     if (row.id) {
-      this.store.delete(row.id).subscribe({
-        next: () => {
-          this.snackbar.open('Departamento eliminado exitosamente', 'Cerrar', { duration: 2000 });
+      this.sweetAlert.showConfirm('¿Está seguro?', 'Esta acción no se puede deshacer').then(result => {
+        if (result.isConfirmed) {
+          this.store.delete(row.id).subscribe({
+            next: () => {
+              this.sweetAlert.showNotification('Éxito', 'Departamento eliminado exitosamente', 'success');
+            }
+          });
         }
       });
     }
